+++ src/melodia/__init__.py
@@ -3,9 +3,11 @@ from typing import Sequence
 from . import core
 from . import io
 from . import music
+from . import processing
 
 __all__: Sequence[str] = [
     'core',
     'io',
-    'music'
+    'music',
+    'processing'
 ]